IFS-261: Fix barcode scanning of wrong file (#7)
JIRA Issue: [IFS-261](https://agile.8x8.com/browse/IFS-271)

# Bug Description
The barcode detector service is scanning the wrong file. This happens
because it uses the `file` field of the kafka event to know which file
it should scan.
Because fax server puts in this field the generated pdf containing the
fax, the barcode service is scanning the wrong file.

# Fix Description
The fix is simple. The original TIF file uses the same name which the
generated pdf file is using. The only thing that changes is the
extension. As so, all it takes is to substitute PDF for TIF.

# Test Description
Test of the fix was done locally by sending a kafka event where the file
field contains the file document name using the PDF extension. Barcode
service was validated by watching the produced kafka event and verifying
that it contained the barcodes of the associated TIF file. Logs were
also verified to check if the extension substitution had been done and
if the name deduction had been done properly.

Signed-off-by: Carlos Miguel Ferreira <carlos.ferreira@8x8.com>

Signed-off-by: Carlos Miguel Ferreira <carlos.ferreira@8x8.com>