IFS-375: Prevent logging database password (#28)
JIRA Issue: IFS-375 [1]

# Task Description
Barcode service is currently logging the database password. This is not
acceptable and this commit addresses this issue.

# Implementation Description
A temporary structure is created with the configuration gathered from
the selected source. Then, the password information is removed and the
rest of the information is logged.

# Test Description
Tested locally using Cloud8 Spring Config service

[1]: https://agile.8x8.com/browse/IFS-375

Signed-off-by: Carlos Ferreira <carlos.ferreira@8x8.com>