IFS-305: Substitute OpenCV for Pillow module (#27)
JIRA Issue: IFS-305 [1]

The use of OpenCV is limited to only loading TIF images into memory.
Unfortunately, `opencv-python` is quite a big library and contains lots
of dependencies required for unused features.
As so, an investigation was done to find a simpler substitute that is
compatible with PyZbar. The Pillow module was the chosen substitute. The
substitution also allows to substantially reduce the size of the set of
dependencies, were reduction was seen to be in the dozens of MBytes.
This is important for the containerization of the service.

The code update is small and contained to the zbar reader module of the
source code. The use of `opencv` was substituted with `Pillow`, where
the rest of the code remained unchanged.
Dependencies manifest was also updated, along with the removal of the
`cv2` module from the pylint whitelist.

This code update was tested by testing the barcode scanner code with
several TIF images which the Fax team has in storage for this effect.
All barcodes that were expected to be detected and decoded were so.

[1]: https://agile.8x8.com/browse/IFS-305

Signed-off-by: Carlos Miguel Ferreira <carlos.ferreira@8x8.com>