IFS-282: Improve detection of available char length for barcode storage (#15)
JIRA Issue: [IFS-282](https://agile.8x8.com/browse/IFS-282)

# Task Description
The objective of the task is to detect the maximum character storage of
the `barCodes` column at the database.
Until now, the barcode service would detect the type of the column and
by knowing it, it would set the maximum available length. This however
had a flaw, as it would not work with `varchar` column types. As so, an
improved way of detecting the maximum amount of characters is now set
with this PR.

# Implementation Description
A SQL Query was devised to obtain the maximum storage length of the
`barCodes` column. This query substitutes the previous one of detecting
the column type.

# Testing Description
Two tests were conducted:
- Service was started against the testing database which uses the same
column type which the database from Pilot is using. Length was
successfully detected.
- `barCodes` column type was updated to `Varchar(50000)` at the testing
database. Length was successfully detected.

Signed-off-by: Carlos Miguel Ferreira <carlos.ferreira@8x8.com>

---------

Signed-off-by: Carlos Miguel Ferreira <carlos.ferreira@8x8.com>
Signed-off-by: Carlos Ferreira <carlos.ferreira@8x8.com>