IFS-404: Implement Health endpoint for readiness and liveness probing (#33)
JIRA Issue: IFS-404 [1]

# Task Description
In order for K8s to properly manage a service, it needs to probe the
service for it's readiness and liveness. As so, barcode service REST API
needs a health endpoint.

# Implementation Description
A new endpoint is implemented at /health that simply returns a 200OK
Given the simplicity of the barcode service REST API, returning an
answer is prove enough that the service is ready to receive requests.

# Testing Description
Locally tested by calling the endpoint. Everything checks out.

[1]: https://agile.8x8.com/browse/IFS-404

Signed-off-by: Carlos Ferreira <carlos.ferreira@8x8.com>