IFS-456: Improve barcode detection using Gaussian Blur filter (#35)
JIRA Issue: IFS-456 [1]

# Task Description
Adapt Health has provided feedback that some barcodes were not being
decoded even though the fax quality was acceptable. This issue was
tracked to artefacts caused by the dithering filter that is applied by
fax machines to reduce the fax size prior to sending.

After doing some analysis and experimentation with Gaussian Blur
filters, it was possible to clear the dithering artefacts and improve
the barcodes scanning process.

# Implementation Description
Gaussian Blue filters will now be used to pre-process the image prior to
scanning it. In order to prevent barcode scanning failures after
applying the filtering, the original image will still be scanned and
barcodes will be extracted.
After scanning the original image and the filtered image, the results
are compared and for barcodes equally returned by both images, the
barcode with higher quality will be the one selected.

# Test Description
In order to make sure that there are no regressions in the future if
more improvements are done to the scanning process, the file provided by
Adapt Health was added to the unit test that validates the scanner.

[1]: https://agile.8x8.com/browse/IFS-456

---------

Signed-off-by: Carlos Ferreira <carlos.ferreira@8x8.com>